manage admin users permissions

+++ app/Http/Controllers/Admin/PadFieldController.php
@@ -9,6 +9,8 @@ class PadFieldController extends Controller
 {
     public function destroy(PadField $padField)
     {
+        abort_if(authUser()->cannot('Manage Admin Panel Pads'), 403);
+
         abort_if($padField->pad->isEnabled(), 403);
 
         $padField->forceDelete();
